requirements.txt
python-telegram-bot
google-genai
PyMuPDF
pymongo
